Solution to Categorize Students Based on Grades

+++ Categorize Students Based on Grades@@67877d61dc107486ad5c274e/code.c
@@ -2,14 +2,14 @@
 struct student{
     int roll_number;
     char name[100];
-    float marks;
+    int marks;
 };
 int main(){
     int n;
     scanf("%d",n);
     struct student s1[n];
     for(int i=0;i<n;i++){
-        scanf("%d %s %.1f",&s1[i].roll_number,&s1[i].name,&s1[i].marks);
+        scanf("%d %s %d",&s1[i].roll_number,&s1[i].name,&s1[i].marks);
     }
     for(int i=0;i<n;i++){
         if(s1[i].marks>=85){
